Add migrations to new model for record entries
We do not provide a method to migrate down without losing data because
it would probably be far more effort than it is worth. I had to manually
add migrations from the old JSON format to the new table. Existing data
will not be cleared until the user re-saves their character (to be added
in a later commit).